added support for nutrition and health categories

+++ projects/src/main/java/com/debjyotipaul/util/DataLoader.java
@@ -122,7 +122,7 @@ public class DataLoader {
                             new Tweet(375,Double.parseDouble(tokens[4]),Double.parseDouble(tokens[5]),tokens[2],Long.parseLong(tokens[0]),
                                     "",
                                     "http://www.panoramio.com/user/475995","http://mw2.google.com/mw-panoramio/photos/medium/11630238.jpg",
-                                    "http://www.panoramio.com/photo/11630238",upload_date,500,Integer.parseInt(tokens[1]));
+                                    "http://www.panoramio.com/photo/11630238",upload_date,500,Long.parseLong(tokens[1]));
 
                     List<String> categories = new ArrayList<String>();
                     categories.add(tokens[11]);
